requirements.txt
flask
mysql-connector-python
pandas
mftool
matplotlib
scikit-learn
numpy
